【产研系统-用户调研-9月优化上】项目文档

+++ dal/src/main/java/com/timevale/forward/dal/dao/ProjectDocumentMapper.java
@@ -11,7 +11,7 @@ import org.apache.ibatis.annotations.Param;
  * @date 2022/8/31/031 18:11
  */
 public interface ProjectDocumentMapper {
-    int insert(ProjectDocument record);
+    Long insert(ProjectDocument record);
 
     int insertSelective(ProjectDocument record);
